design/스크롤 영역 수정
