+++ src/screens/Letter/LetterHome/index.tsx
@@ -1,25 +1,25 @@
 import {getAlarmCategory} from '@apis/alarm';
 import {getLetters} from '@apis/providedFile';
-import EmptyLetterIcon from '@assets/svgs/emptyLetter.svg';
 import AnimatedView from '@components/atom/AnimatedView';
 import BottomMenu from '@components/atom/BottomMenu';
 import Button from '@components/atom/Button';
 import Modal from '@components/atom/Modal';
 import Toast from '@components/atom/Toast';
 import Txt from '@components/atom/Txt';
-import {EMOTION_OPTIONS, LETTERS_DATA} from '@constants/letter';
+import VirtualizedView from '@components/atom/VirtualizedView';
+import {LETTERS_DATA} from '@constants/letter';
 import {Portal} from '@gorhom/portal';
 import useGetSummary from '@hooks/providedFile/useGetSummary';
 import usePostReport from '@hooks/providedFile/usePostReport';
 import useModal from '@hooks/useModal';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import LetterCard from '@screens/Letter/LetterList/components/LetterCard';
+import ListArea from '@screens/Letter/LetterHome/components/ListArea';
+import ListHeader from '@screens/Letter/LetterHome/components/ListHeader';
 import {LetterResponseData} from '@type/api/providedFile';
 import {LetterStackParamList} from '@type/nav/LetterStackParamList';
 import {useEffect, useState} from 'react';
-import {Alert, Pressable, Text, View} from 'react-native';
-import {ScrollView} from 'react-native-gesture-handler';
+import {Alert, Pressable, View} from 'react-native';
 
 type LetterProps = NativeStackScreenProps<
   LetterStackParamList,
@@ -111,10 +111,6 @@ const LetterHomeScreen = ({navigation}: Readonly<LetterProps>) => {
     setFilteredLettersData(filteredLetters);
   }, [selectedFilterIdx]);
 
-  const displayNum = (num: number) => {
-    return num > 999 ? '999+' : String(num);
-  };
-
   const [selectedFileId, setSelectedFileId] = useState<number | null>(null);
   const [isToast, setIsToast] = useState(false); // 토스트 메시지 표시 상태
   const [toastMessage, setToastMessage] = useState(''); // 토스트 메시지
@@ -138,164 +134,29 @@ const LetterHomeScreen = ({navigation}: Readonly<LetterProps>) => {
 
   return (
     <View className="bg-blue700 flex-1">
-      <View className="px-[30]">
-        <View className="h-[61]" />
-        <Txt
-          type="title2"
-          text={`${nickname}님의 목소리를`}
-          className="text-white"
-        />
-        <View className="flex-row">
-          <View className="flex-row justify-start items-center">
-            <Txt type="title2" text="청년들이 " className="text-white" />
-            <Txt
-              type="title2"
-              text={`${String(summaryData?.result.totalListeners ?? '')}번`}
-              className="text-yellowPrimary"
-            />
-            <Txt type="title2" text=" 청취했어요" className="text-white" />
-          </View>
-        </View>
-
-        <View className="h-[31]" />
-
-        <View className="flex-row items-center">
-          {EMOTION_OPTIONS.slice(0, 2).map(emotion => (
-            <View key={emotion.label} className="flex-row items-center flex-1">
-              <View className="flex-row items-center flex-1">
-                {emotion.icon}
-                <View className="w-[5]" />
-                <Txt type="body3" text={emotion.label} className="text-white" />
-                <View className="w-[8]" />
-                <Txt
-                  type="body3"
-                  text={displayNum(
-                    summaryData?.result.reactionsNum[
-                      emotion.type as keyof typeof summaryData.result.reactionsNum
-                    ] ?? 0,
-                  )}
-                  className="text-yellowPrimary"
-                />
-              </View>
-              <View className="w-[41]" />
-            </View>
-          ))}
-        </View>
-        <View className="h-[15]" />
-        <View className="flex-row items-center">
-          {EMOTION_OPTIONS.slice(2).map(emotion => (
-            <View key={emotion.label} className="flex-row items-center flex-1">
-              <View className="flex-row items-center flex-1">
-                {emotion.icon}
-                <View className="w-[5]" />
-                <Txt type="body3" text={emotion.label} className="text-white" />
-                <View className="w-[8]" />
-                <Txt
-                  type="body3"
-                  text={String(
-                    summaryData?.result.reactionsNum[
-                      emotion.type as keyof typeof summaryData.result.reactionsNum
-                    ] ?? '',
-                  )}
-                  className="text-yellowPrimary"
-                />
-              </View>
-              <View className="w-[41]" />
-            </View>
-          ))}
-        </View>
-      </View>
-
-      <View className="h-[45]" />
-
-      <View className="bg-blue600 flex-1">
-        {/* 편지 카테고리 */}
-        <View>
-          <ScrollView
-            horizontal
-            showsHorizontalScrollIndicator={false}
-            className="py-[30]">
-            <View className="flex-row items-center px-[30] h-[36]">
-              {parentCategories.map((menu, index) => (
-                <Pressable
-                  key={`${menu.category}=${menu.label}-parentCategories`}
-                  className={`h-[36] px-[22] items-center justify-center border ${
-                    index === selectedFilterIdx
-                      ? 'border-blue400 bg-white/10'
-                      : 'border-white10'
-                  } mr-[8]`}
-                  style={{borderRadius: 20}}
-                  onPress={() => setSelectedFilterIdx(index)}>
-                  <Txt
-                    type="body4"
-                    text={menu.label}
-                    className={`${
-                      index === selectedFilterIdx
-                        ? 'text-yellowPrimary'
-                        : 'text-gray300'
-                    }`}
-                  />
-                </Pressable>
-              ))}
-            </View>
-          </ScrollView>
-        </View>
-
+      {/* <ScrollView stickyHeaderIndices={[2]}> */}
+      <VirtualizedView>
+        <ListHeader nickname={nickname} summaryData={summaryData} />
         {/* 편지 리스트 */}
-        <View className="px-[30] pb-[280]">
-          <View className="flex-row items-center">
-            <Text
-              className="text-white"
-              style={{
-                fontSize: 22,
-                fontFamily: 'LeeSeoyun-Regular',
-                lineHeight: 22 * 1.5,
-              }}>
-              TO.
-            </Text>
-            <Txt
-              type="title4"
-              text={nickname ?? ''}
-              className="text-yellowPrimary ml-[7]"
-            />
-          </View>
-
-          <View className="h-[22]" />
-
-          <View className="">
-            {!filteredLettersData || filteredLettersData.length === 0 ? (
-              <View className="items-center justify-center pt-[40]">
-                <EmptyLetterIcon />
-                <View className="h-[26.27]" />
-                <Txt
-                  type="caption1"
-                  text={`아직 받은 편지가 없어요\n청년들에게 더 많은 목소리를 전해보세요`}
-                  className="text-blue300 text-center"
-                />
-              </View>
-            ) : (
-              <ScrollView className="">
-                <View className="">
-                  {filteredLettersData.map((letter, idx) => (
-                    <View
-                      key={`${letter.createdAt}-${letter.providedFileId}-filteredLettersData`}>
-                      <LetterCard
-                        letter={letter}
-                        idx={idx}
-                        onPressMoreDot={() => {
-                          setClickedMoreDot(true);
-                          setSelectedFileId(letter.providedFileId);
-                        }}
-                      />
-                      <View className="mb-[30]" />
-                    </View>
-                  ))}
-                </View>
-              </ScrollView>
-            )}
-          </View>
-        </View>
-      </View>
+        <ListArea
+          nickname={nickname}
+          list={filteredLettersData}
+          // list={LETTERS_DATA}
+          setClickedMoreDot={setClickedMoreDot}
+          setSelectedFileId={setSelectedFileId}
+          selectedFilterIdx={selectedFilterIdx}
+          setSelectedFilterIdx={setSelectedFilterIdx}
+          parentCategories={parentCategories}
+          // parentCategories={[
+          //   {category: 'WAKE_UP1', label: '기상'},
+          //   {category: 'WAKE_UP2', label: '기상'},
+          //   {category: 'WAKE_UP3', label: '기상'},
+          //   {category: 'WAKE_UP4', label: '기상'},
+          //   {category: 'SLEEP', label: '취침'},
+          // ]}
+        />
+        {/* </ScrollView> */}
+      </VirtualizedView>
 
       <Portal>
         <Pressable
